requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
tflite-runtime
